feat(darwin): use only one machine for now
Previously, the default app determined the flake attribute to run from
the hostname of the host machine. This caused issues in CI because no
flake attribute was defined for the GitHub Actions macOS runner.

For now, we will simply use the system name as the flake attribute name.
Later, we can make it flexible to allow multiple machine configurations.